change some values to use float instead of int

+++ src/agi-eval.ts
@@ -1,42 +1,60 @@
 export function calcAgility() {
   let charBaseAgiNode = document.getElementById("cba") as HTMLInputElement;
-  let charTimedAgiPassiveNode = document.getElementById("ctap") as HTMLInputElement;
-  let charRankFlatTotalNode = document.getElementById("cfr") as HTMLInputElement;
-  let charRankPCTTotalNode = document.getElementById("cpctr") as HTMLInputElement;
+  let charTimedAgiPassiveNode = document.getElementById(
+    "ctap"
+  ) as HTMLInputElement;
+  let charRankFlatTotalNode = document.getElementById(
+    "cfr"
+  ) as HTMLInputElement;
+  let charRankPCTTotalNode = document.getElementById(
+    "cpctr"
+  ) as HTMLInputElement;
   let flatMemory1AgiNode = document.getElementById("m1f") as HTMLInputElement;
   let flatMemory2AgiNode = document.getElementById("m2f") as HTMLInputElement;
   let flatMemory3AgiNode = document.getElementById("m3f") as HTMLInputElement;
   let pctMemoryAgiNode = document.getElementById("mpct") as HTMLInputElement;
-  let weapon1FlatAgiPassiveNode = document.getElementById("w1f") as HTMLInputElement;
-  let weapon1TimedAgiPassiveNode = document.getElementById("w1pct") as HTMLInputElement;
-  let weapon2FlatAgiPassiveNode = document.getElementById("w2f") as HTMLInputElement;
-  let weapon2TimedAgiPassiveNode = document.getElementById("w2pct") as HTMLInputElement;
-  let weapon3FlatAgiPassiveNode = document.getElementById("w3f") as HTMLInputElement;
-  let weapon3TimedAgiPassiveNode = document.getElementById("w3pct") as HTMLInputElement;
+  let weapon1FlatAgiPassiveNode = document.getElementById(
+    "w1f"
+  ) as HTMLInputElement;
+  let weapon1TimedAgiPassiveNode = document.getElementById(
+    "w1pct"
+  ) as HTMLInputElement;
+  let weapon2FlatAgiPassiveNode = document.getElementById(
+    "w2f"
+  ) as HTMLInputElement;
+  let weapon2TimedAgiPassiveNode = document.getElementById(
+    "w2pct"
+  ) as HTMLInputElement;
+  let weapon3FlatAgiPassiveNode = document.getElementById(
+    "w3f"
+  ) as HTMLInputElement;
+  let weapon3TimedAgiPassiveNode = document.getElementById(
+    "w3pct"
+  ) as HTMLInputElement;
 
   let charBaseAgi = parseInt(charBaseAgiNode.value);
-  let charTimedAgiPassive = parseInt(charTimedAgiPassiveNode.value) / 100;
+  let charTimedAgiPassive = parseFloat(charTimedAgiPassiveNode.value) / 100;
 
   let charRankFlatTotal = parseInt(charRankFlatTotalNode.value);
-  let charRankPCTTotal = parseInt(charRankPCTTotalNode.value) / 100;
+  let charRankPCTTotal = parseFloat(charRankPCTTotalNode.value) / 100;
 
   let flatMemory1Agi = parseInt(flatMemory1AgiNode.value);
   let flatMemory2Agi = parseInt(flatMemory2AgiNode.value);
   let flatMemory3Agi = parseInt(flatMemory3AgiNode.value);
 
-  let pctMemoryAgi = parseInt(pctMemoryAgiNode.value) / 100;
+  let pctMemoryAgi = parseFloat(pctMemoryAgiNode.value) / 100;
 
-  let weapon1FlatAgiPassive = parseInt(weapon1FlatAgiPassiveNode.value) / 100;
+  let weapon1FlatAgiPassive = parseFloat(weapon1FlatAgiPassiveNode.value) / 100;
   let weapon1TimedAgiPassive =
-    parseInt(weapon1TimedAgiPassiveNode.value) / 100;
+    parseFloat(weapon1TimedAgiPassiveNode.value) / 100;
 
-  let weapon2FlatAgiPassive = parseInt(weapon2FlatAgiPassiveNode.value) / 100;
+  let weapon2FlatAgiPassive = parseFloat(weapon2FlatAgiPassiveNode.value) / 100;
   let weapon2TimedAgiPassive =
-    parseInt(weapon2TimedAgiPassiveNode.value) / 100;
+    parseFloat(weapon2TimedAgiPassiveNode.value) / 100;
 
-  let weapon3FlatAgiPassive = parseInt(weapon3FlatAgiPassiveNode.value) / 100;
+  let weapon3FlatAgiPassive = parseFloat(weapon3FlatAgiPassiveNode.value) / 100;
   let weapon3TimedAgiPassive =
-    parseInt(weapon3TimedAgiPassiveNode.value) / 100;
+    parseFloat(weapon3TimedAgiPassiveNode.value) / 100;
 
   const charRankAgi = charRankFlatTotal + charBaseAgi * charRankPCTTotal;
   const totalMemoryAgi =
